chore(contracts): don't export deploy script by default
