+++ packages/contracts/tsconfig.json
@@ -3,6 +3,6 @@
     "compilerOptions": {
         "outDir": "./build"
     },
-    "include": ["./src", "./typechain"],
+    "include": ["./src", "./deploy", "./typechain"],
     "files": ["./hardhat.config.ts"]
 }
