+++ packages/core/test/UserState/UserState.test.ts
@@ -1,7 +1,7 @@
 import { ethers } from 'hardhat'
 import { expect } from 'chai'
 import { ZkIdentity, hashLeftRight } from '@unirep/crypto'
-import { deployUnirep } from '@unirep/contracts'
+import { deployUnirep } from '@unirep/contracts/deploy'
 import {
     EPOCH_LENGTH,
     EPOCH_TREE_DEPTH,
