+++ packages/core/test/UserState/userSignUpProofEvent.test.ts
@@ -3,7 +3,8 @@ import { ethers as hardhatEthers } from 'hardhat'
 import { ethers } from 'ethers'
 import { expect } from 'chai'
 import { ZkIdentity, hashLeftRight } from '@unirep/crypto'
-import { deployUnirep, Unirep } from '@unirep/contracts'
+import { Unirep } from '@unirep/contracts'
+import { deployUnirep } from '@unirep/contracts/deploy'
 
 import { Reputation, computeInitUserStateRoot } from '../../src'
 import { genUserState, genUnirepState, genNewGST } from '../utils'
