+++ packages/core/test/synchronizer/synchronizer.test.ts
@@ -4,7 +4,7 @@ import { expect } from 'chai'
 import { ZkIdentity, hashLeftRight } from '@unirep/crypto'
 import { EPOCH_LENGTH } from '@unirep/circuits'
 import { defaultProver } from '@unirep/circuits/provers/defaultProver'
-import { deployUnirep } from '@unirep/contracts'
+import { deployUnirep } from '@unirep/contracts/deploy'
 
 const attestingFee = ethers.utils.parseEther('0.1')
